requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
fpdf==1.7.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
packaging==24.1
pillow==11.0.0
python-telegram-bot==21.6
requests==2.32.3
sniffio==1.3.1
tornado==6.4.1
urllib3==2.2.3
